feat: 예약상세 핸디 제거 및 공지방참여 섹션, 탑승권 섹션 추가
