+++ src/app/mypage/shuttle/reservation/[reservationId]/components/Content.tsx
@@ -4,7 +4,6 @@ import { PaymentsViewEntity } from '@/types/payment.type';
 import { ReservationsViewEntity } from '@/types/reservation.type';
 import { ShuttleBusesViewEntity } from '@/types/shuttleBus.type';
 import EventCard from './EventCard';
-import ShuttleProgressSection from './sections/shuttle-progress-section/ShuttleProgressSection';
 import useReservationProgress, {
   ReservationProgress,
 } from '../../../hooks/useReservationProgress';
@@ -18,6 +17,8 @@ import PrimaryCheckIcon from '../icons/icon-check-primary.svg';
 import GreyCheckIcon from '../icons/icon-check-grey.svg';
 import WrapperWithDivider from './WrapperWithDivider';
 import { checkIsHandyParty } from '@/utils/handyParty.util';
+import HandyPartyProgressSection from './sections/shuttle-progress-section/HandyPartyProgressSection';
+import Button from '@/components/buttons/button/Button';
 
 interface Props {
   reservation: ReservationsViewEntity;
@@ -60,6 +61,13 @@ const Content = ({ reservation, payment, shuttleBus }: Props) => {
 
   const isHandyParty = checkIsHandyParty(shuttleRoute);
 
+  const noticeRoomUrl = dailyEvent?.metadata?.openChatUrl;
+  const handleOpenNoticeRoom = () => {
+    if (noticeRoomUrl) {
+      window.open(noticeRoomUrl, '_blank', 'noopener,noreferrer');
+    }
+  };
+
   return (
     <main className="grow pb-16">
       {isHandyParty && (
@@ -71,18 +79,48 @@ const Content = ({ reservation, payment, shuttleBus }: Props) => {
       <Title progress={reservationProgress} />
       <EventCard event={event} />
       <ul className="flex flex-col gap-24">
-        {!isCanceled && (
-          <WrapperWithDivider>
-            <ShuttleProgressSection
-              reservation={reservation}
-              reservationProgress={reservationProgress}
-              isOpenChatLinkCreated={isOpenChatLinkCreated}
-              handyStatus={handyStatus}
-              shuttleBus={shuttleBus}
-              isHandyParty={isHandyParty}
-            />
-          </WrapperWithDivider>
-        )}
+        {!isCanceled &&
+          (isHandyParty ? (
+            <WrapperWithDivider>
+              <HandyPartyProgressSection
+                reservationProgress={reservationProgress}
+                isOpenChatLinkCreated={isOpenChatLinkCreated}
+                handyStatus={handyStatus}
+                shuttleBus={shuttleBus}
+                isHandyParty={isHandyParty}
+              />
+            </WrapperWithDivider>
+          ) : (
+            <>
+              <WrapperWithDivider>
+                <section className="flex flex-col gap-16 px-16">
+                  <Button>탑승권 확인하기</Button>
+                  <p className="text-14 font-500 leading-[160%]">
+                    현장에서 탑승권을 제시한 후 탑승해 주세요.{' '}
+                    <span className="font-600 text-basic-red-400">
+                      캡쳐 이미지로는 탑승이 제한될 수 있는 점 참고해 주세요.
+                    </span>
+                  </p>
+                </section>
+              </WrapperWithDivider>
+              <WrapperWithDivider>
+                {/* NOTE: 일자별 노선에 오픈채팅방이 먼저 반영되고 출시되어야 하는 기능 */}
+                <section className="flex flex-col gap-16 px-16">
+                  <Button
+                    variant="secondary"
+                    onClick={handleOpenNoticeRoom}
+                    disabled={!noticeRoomUrl}
+                  >
+                    공지방 참여하기
+                  </Button>
+                  <p className="text-14 font-500 leading-[160%]">
+                    탑승 당일, 셔틀버스 변동사항 및 실시간 안내사항은 공지방에서
+                    이루어져요. 탑승 전 반드시 참여해 주세요!
+                  </p>
+                </section>
+              </WrapperWithDivider>
+            </>
+          ))}
         <WrapperWithDivider>
           <ShuttleInfoSection
             tripType={reservation.type}
@@ -110,12 +148,7 @@ const Content = ({ reservation, payment, shuttleBus }: Props) => {
         </WrapperWithDivider>
         {!isHandyParty && (
           <WrapperWithDivider>
-            <HandySection
-              isCanceled={isCanceled}
-              isEnded={isEnded}
-              handyStatus={handyStatus}
-              reservationId={reservation.reservationId}
-            />
+            <HandySection handyStatus={handyStatus} />
           </WrapperWithDivider>
         )}
         <WrapperWithDivider>
@@ -123,7 +156,6 @@ const Content = ({ reservation, payment, shuttleBus }: Props) => {
             payment={payment}
             passengerCount={reservation.passengerCount}
             isCanceled={isCanceled}
-            isHandySupported={handyStatus !== 'NOT_SUPPORTED'}
           />
         </WrapperWithDivider>
         <GuidelineSection />
