+++ src/app/mypage/shuttle/reservation/[reservationId]/components/sections/price-section/components/RegularPriceContent.tsx
@@ -6,14 +6,9 @@ import RefundRequestList from './RefundRequestList';
 interface Props {
   payment: PaymentsViewEntity;
   passengerCount: number;
-  isHandySupported: boolean;
 }
 
-const RegularPriceContent = ({
-  payment,
-  passengerCount,
-  isHandySupported,
-}: Props) => {
+const RegularPriceContent = ({ payment, passengerCount }: Props) => {
   const paymentAmount = payment.paymentAmount;
   const regularPrice = payment.principalAmount / passengerCount;
   const totalEarlybirdDiscountAmount = payment.earlybirdDiscountAmount;
@@ -79,13 +74,6 @@ const RegularPriceContent = ({
           결제 일시 | {paymentAt}
         </span>
       </li>
-      {isHandySupported && (
-        <div className="rounded-8 bg-basic-grey-50 p-8 text-12 font-500 text-basic-grey-500">
-          핸디 지원금은 결제 금액의 50%이며, 선정된 핸디에 한해 셔틀 종료 후
-          환급됩니다. (영업일 기준 3일 이내) 결제 단계에서 할인이 적용되지
-          않습니다.
-        </div>
-      )}
       <RefundRequestList refundRequests={payment.refundRequests} />
     </div>
   );
